fix: поправь замечания ревьюера (отображение чужих сохранённых фильмов) и вынеси в отдельный хук useEscapePress
